\name{fitsoilwater_App}
\alias{fitsoilwater_App}

\title{A shiny for fitting soil water retention curves}

\description{A shiny for fitting soil water retention curves}

\usage{
fitsoilwater_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}