\name{HEMC_App}
\alias{HEMC_App}

\title{A shiny for fitting High Energy Moisture Characteristic (HEMC) curve for measuring aggregate stability}

\description{A shiny for fitting High Energy Moisture Characteristic (HEMC) curve for measuring aggregate stability}

\usage{
HEMC_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
