\name{LLWR_App}
\alias{LLWR_App}

\title{A shiny for calculation of the usual least limiting water range}

\description{A shiny for calculation of the usual least limiting water range}

\usage{
LLWR_App()
}

\value{A shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
