\name{particle.sedimentation_App}
\alias{particle.sedimentation_App}

\title{A interactive web page for time of particle sedimentation}

\description{A interactive web page for time of particle sedimentation}

\usage{
particle.sedimentation_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{Time of particle sedimentation}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
