\name{PredComp}
\alias{PredComp}

\title{A shiny for simulation of soil compaction}

\description{A shiny for for simulation of soil compaction}

\usage{
PredComp()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
