\name{SoilAggregate}
\alias{SoilAggregate}
\docType{data}

\title{Soil Aggregate Size Data Set}

\description{
Data set for determining soil aggregate size distribution.
}

\usage{data(SoilAggregate)}

\format{
  A data frame with 12 observations on 7 variables.
  \describe{
    \item{\code{ID}}{a factor with the names of the soil samples.}
    \item{\code{D3}}{}
    \item{\code{D1.5}}{}
    \item{\code{D0.75}}{}
    \item{\code{D0.375}}{}
    \item{\code{D0.178}}{}
    \item{\code{D0.053}}{}
  }
}

\examples{
data(SoilAggregate)
summary(SoilAggregate)
}

\keyword{datasets}

