\name{WRC_App}
\alias{WRC_App}

\title{A shiny to automatically fit the water retention curve}

\description{Use the maximum likelihood method to find the water retention curve that best fits the data.}

\usage{
WRC_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{shiny app}

\author{ Anderson Rodrigo da Silva <anderson.agro@hotmail.com> }

\seealso{
  \code{\link{fitsoilwater_App}}
}
