\name{aggreg.stability_App}
\alias{aggreg.stability_App}

\title{A shiny for Soil Aggregate-Size Distribution}

\description{A shiny for Soil Aggregate-Size Distribution}

\usage{
aggreg.stability_App()
}

\value{A shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
