\name{LLWR_LLMPRApp}
\alias{LLWR_LLMPRApp}

\title{A shiny APP for LLWR and LLMPR calculation}

\description{A shiny APP for LLWR and LLMPR calculation}

\usage{
LLWR_LLMPRApp()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
