\name{LLWR_LLMPR_App}
\alias{LLWR_LLMPR_App}

\title{A shiny APP for LLWR and LLMPR calculation}

\description{A shiny APP for LLWR and LLMPR calculation}

\usage{
LLWR_LLMPR_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
