\name{LLWR_LLMPR_App}
\alias{LLWR_LLMPR_App}

\title{A shiny for calculation of least limiting water and matric potential ranges of agricultural soils with calculated physical restriction thresholds}

\description{A shiny for calculation of least limiting water and matric potential ranges of agricultural soils with calculated physical restriction thresholds}

\usage{
LLWR_LLMPR_App()
}

\value{A shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
