\name{LLWR_LLMPR_shiny}
\alias{LLWR_LLMPR_shiny}

\title{A shiny APP for LLWR and LLMPR calculation}

\description{A shiny APP for LLWR and LLMPR calculation}
\usage{
LLWR_LLMPR_shiny()
}
\arguments{
  \item{LLWR_LLMPR_shiny}{Available on the interactive web page}

}
\value{LLWR and LLMPR}
\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\references{

}


