\name{particle.sedimentation_App}
\alias{particle.sedimentation_App}

\title{A shiny for time of particle sedimentation}

\description{A shiny for time of particle sedimentation}

\usage{
particle.sedimentation_App()
}

\value{A shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
