\name{r_App}
\alias{r_App}

\title{A interactive web page for equation of capillarity}

\description{A interactive web page for equation of capillarity}

\usage{
r_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{Pores that drain for a corresponding applied matric suction}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
