\name{r_App}
\alias{r_App}

\title{A shiny for equation of capillarity}

\description{A shiny for equation of capillarity}

\usage{
r_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

\seealso{
  \code{\link{stressTraffic}}
}
